chore: testing on-demand on fit
